Create To check wheathwer the given char present in the string
Write a program that finds whether a given character is present in a string or not. In case it is present it prints the index at which it is present. Do not use built-in find functions to search the character.
Sample Input:
	Enter the string: I am a programmer
	Enter the character to be searched: p
Sample Output:
	P is found in string at index: 8
Note: Check for non available Character in the given statement as Hidden Test case.